logs/requirements.txt
psycopg2-binary==2.9.3
Flask==2.3.3
requests==2.31.0
flask-cors
